0.0.5 BUG FIX
Fixed a bug in the construction of the Hopping matrix for Lattice Kondo.

BUG: the algorithm was naively using the number arrays Nup=[Nele_up, Nimp_up] and
Ndw=[Nele_dw,Nimp_dw] to check the conditions to applye the chain of operators c.cdg
After a simple examination I realized that this choice is obviously wrong as the
Hamiltonian Hij(io,jo) and the numbering of the ket follows the binary decomposition 1:Ns;1:Ns.

FIX: Using the binary decomposed array Ib and indiced io,jo -> io+Ns,jo+Ns for spins Up and Dw.