0.0.6 FINAL BUG FIX
In the effort of treating the hopping part of H in a homogeneous way I
introduced a significant bug, which is now (hopefully) fixed.

BUG: the hop matrix is implemented with this pseudo-code

for m=1,Nstates
do
for i,j=1:Ns
do
   if H(i,j)!=0 && nup(i)=1 && nup(j)=0 ; then
    apply C_up,map(i) Cdg_up,map(j) |m>
end
end

where map(i) should map the label i to the correct position in the ket structure:
|1...eNs>_up |1...eNs>_dw |1...iNs>_up |1...iNs>_dw
 1,.,eNs,eNs+1,..,2eNs,2eNs+1,..,2eNs+iNs,...Ns
the map send the indices i=1,..Ns to the corresponding correct position in the bit-decomposition of the state,
i.e. map_up [1,...,Ns]--> [1...eNs,2eNs+1,..,2eNs+iNs]

The map used in the previous implementations was wrong and caused a number of inconsistencies
with application of opeartors C or Cdg resulting in zero

FIX:
constructed the correct map *ket_index* in AUX_FUNX